.github/copilot/demo-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
mysql-connector-python==8.3.0
httpx==0.26.0
